1.修复初次打开播放没有声音问题 2.修复界面标题栏部分遮挡问题

+++ video_player_core/module/sdl/core_sdl_op.h
@@ -41,7 +41,7 @@ public:
     bool initResample(AVCodecContext* ctx);
     bool init(audioCallback cb, void* userdata);
     void setVol(int vol);
-    float getVol();
+    int getVol();
 
     void resampleFrame(AVFrame*, int& bufferSize);
